# spec/kompot.openapi.json
{
  "components": {
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "tacku",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/forms/task/{task}": {
      "get": {
        "operationId": "taskScreen",
        "parameters": [
          {
            "in": "path",
            "name": "task",
            "required": true,
            "schema": {
              "pattern": "^TAC-[1-9][0-9]*$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "kompot-forms.schema.json#/$defs/KompotFormResponse"
                }
              }
            },
            "description": "ok"
          },
          "401": {
            "description": "no token, or a token issued for another resource"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-kompot-endpoint-kind": "form"
      }
    },
...
    "/submit/task-view/{task}": {
      "post": {
        "operationId": "submitTaskView",
        "parameters": [
          {
            "in": "path",
            "name": "task",
            "required": true,
            "schema": {
              "pattern": "^TAC-[1-9][0-9]*$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "kompot-core.schema.json#/$defs/KompotAction"
                }
              }
            },
            "description": "ok"
          },
          "400": {
            "description": "no idempotency key"
          },
          "401": {
            "description": "no token, or a token issued for another resource"
          },
          "409": {
            "description": "the key was used for a different request"
          },
          "422": {
            "description": "refused on its merits"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-kompot-endpoint-kind": "submit"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ]
}